store: split tests for refresh/action support to store_action_test.go
